% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectTemplate.R
\docType{class}
\name{TrenaProjectTemplate-class}
\alias{TrenaProjectTemplate-class}
\alias{.TrenaProjectTemplate}
\alias{TrenaProjectTemplate}
\title{TrenaProjectTemplate-class}
\usage{
TrenaProjectTemplate(quiet = TRUE)
}
\value{
An object of the TrenaProjectTemplate class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
